+++ projects/swagger-schematics/mocks/swagger-mock.ts
@@ -2,9 +2,11 @@ import { IPath, ISwaggerSchema } from '../interfaces/version_3_1/swagger.interfa
 import {
     DELETE_MANY_ARRAY_OF_IDS_SWAGGER,
     GET_MODEL_BY_ID_SWAGGER, GET_SERVICE_ACTIONS_BY_ID_SWAGGER, GET_SERVICE_ACTIONS_SWAGGER,
+    GET_BY_STATUS_WITH_ENUM_PARAM_SWAGGER,
     POST_MODEL_CHILD_BY_MODEL_ID_SWAGGER,
     POST_MODEL_FORM_DATA_SWAGGER, POST_SEARCH_ALL_SWAGGER, POST_SEARCH_IDS_SWAGGER,
-    PUT_MODEL_BY_ID_SWAGGER, PUT_MODEL_WITH_EMPTY_BODY_SWAGGER, PUT_MODEL_WITH_INTEGER_BODY_SWAGGER
+    PUT_MODEL_BY_ID_SWAGGER, PUT_MODEL_WITH_EMPTY_BODY_SWAGGER, PUT_MODEL_WITH_INTEGER_BODY_SWAGGER,
+    PUT_WITH_PATH_PARAM_FROM_BODY_SWAGGER
 } from './api-mocks';
 import {API_GET_CHILD_OF_MODEL_BY_ID, ENUM_WITH_VAR_NAMES_SWAGGER, MODEL_WITH_REF_SWAGGER} from './interface-mocks';
 
@@ -50,6 +52,12 @@ export const SWAGGER_DATA: ISwaggerSchema<TMockApiPath> = {
         },
         "/api/Claim/serviceactions/{serviceActionId}": {
             ...GET_SERVICE_ACTIONS_BY_ID_SWAGGER
+        },
+        "/api/Claim/bystatus": {
+            ...GET_BY_STATUS_WITH_ENUM_PARAM_SWAGGER
+        },
+        "/api/Claim/note/{id}": {
+            ...PUT_WITH_PATH_PARAM_FROM_BODY_SWAGGER
         },
     },
     components: {